+++ backend-spring/src/main/java/com/ssafy/pcgg/domain/recommend/repository/RamRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.pcgg.domain.recommend.entity.RamEntity;
 
 @Repository
-public interface RamRepository extends JpaRepository<RamEntity, Integer> {
+public interface RamRepository extends JpaRepository<RamEntity, Long> {
 }
